rename utils to utils.legacy
