+++ src/__tests__/utils.get-a11y-status-message.js
@@ -1,4 +1,4 @@
-import {getA11yStatusMessage} from '../utils'
+import {getA11yStatusMessage} from '../utils.legacy'
 
 const tests = [
   {
